fix: develop 브랜치 마커 구조 통합 및 줌 레벨 마커 관리 개선
- develop 브랜치의 새로운 마커 구조(TeamMarkerFactory, CircuitMarkerManager) 도입
- Legacy 마커 파일들 제거 (TeamMarkerWithLeader.tsx, CircuitMarkerWithLeader.tsx)
- trackManager에 섹터/DRS/Speed Trap 마커 관리 기능 추가
- 줌 레벨 10 미만 시 모든 마커들이 함께 제거되도록 개선
- 마커 등록 시점 문제 해결 (중복 등록 방지)
- ESLint 오류 수정 (unused variables)

주요 개선사항:
- circuitManagerExtensions.ts 추가로 기존 기능 유지
- trackManager 등록 순서 최적화
- 줌 아웃 시 모든 트랙 관련 마커들 완전 제거
